PasarelaPagos/PasarelaDePagos/requirements.txt
stripe>=24.0
python-dotenv>=25.1.1